REFACTOR ISA-10: CalenderView에서 스크롤 제스처 개선 및 MainSheet 재구성
